requirements.txt
python-telegram-bot==22.1
requests
pandas
numpy
mplfinance
pandas-ta
matplotlib
apscheduler
